requirements.txt
asgiref==3.2.3
Django==3.0.2
pytz==2019.3
sqlparse==0.3.0
asgiref==3.2.3
Django==3.0.2
django-js-asset==1.2.2
django-mptt==0.10.0
pytz==2019.3
sqlparse==0.3.0
